# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivityLog": {
        "description": "Activity log schema for API responses.",
        "properties": {
          "container_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container Id"
          },
          "container_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "event_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Metadata"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_type",
          "severity",
          "container_id",
          "container_name",
          "title",
          "description",
          "event_metadata",
          "timestamp",
          "created_at"
        ],
        "title": "ActivityLog",
        "type": "object"
      },
...
      "Container": {
        "description": "Full container schema.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "critical_count": {
            "title": "Critical Count",
            "type": "integer"
          },
          "dive_analyzed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dive Analyzed At"
          },
          "dive_efficiency_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dive Efficiency Score"
          },
          "dive_image_size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dive Image Size Bytes"
          },
          "dive_inefficient_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dive Inefficient Bytes"
          },
          "dive_layer_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dive Layer Count"
          },
          "fixable_vulns": {
            "title": "Fixable Vulns",
            "type": "integer"
          },
          "high_count": {
            "title": "High Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image": {
            "title": "Image",
            "type": "string"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "image_tag": {
            "title": "Image Tag",
            "type": "string"
          },
          "is_my_project": {
            "title": "Is My Project",
            "type": "boolean"
          },
          "is_running": {
            "title": "Is Running",
            "type": "boolean"
          },
          "last_scan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContainerLastScan"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_scan_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scan Date"
          },
          "last_scan_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Scan Status"
          },
          "last_seen": {
            "format": "date-time",
            "title": "Last Seen",
            "type": "string"
          },
          "low_count": {
            "title": "Low Count",
            "type": "integer"
          },
          "medium_count": {
            "title": "Medium Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "scanner_coverage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanner Coverage"
          },
          "total_vulns": {
            "title": "Total Vulns",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "vulnerability_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContainerVulnerabilitySummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "image",
          "image_tag",
          "image_id",
          "id",
          "is_running",
          "is_my_project",
          "last_seen",
          "last_scan_date",
          "last_scan_status",
          "total_vulns",
          "fixable_vulns",
          "critical_count",
          "high_count",
          "medium_count",
          "low_count",
          "scanner_coverage",
          "dive_efficiency_score",
          "dive_inefficient_bytes",
          "dive_image_size_bytes",
          "dive_layer_count",
          "dive_analyzed_at",
          "created_at",
          "updated_at"
        ],
        "title": "Container",
        "type": "object"
      },
...
      "ContainerList": {
        "description": "List of containers with metadata.",
        "properties": {
          "containers": {
            "items": {
              "$ref": "#/components/schemas/ContainerSummary"
            },
            "title": "Containers",
            "type": "array"
          },
          "never_scanned": {
            "title": "Never Scanned",
            "type": "integer"
          },
          "scanned": {
            "title": "Scanned",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "containers",
          "total",
          "scanned",
          "never_scanned"
        ],
        "title": "ContainerList",
        "type": "object"
      },
...
      "ContainerUpdate": {
        "description": "Schema for updating a container.",
        "properties": {
          "is_my_project": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is My Project"
          },
          "is_running": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Running"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          }
        },
        "title": "ContainerUpdate",
        "type": "object"
      },
...
      "ContainerVulnerabilitySummary": {
        "description": "Aggregated vulnerability figures for a container.",
        "properties": {
          "critical": {
            "title": "Critical",
            "type": "integer"
          },
          "fixable": {
            "title": "Fixable",
            "type": "integer"
          },
          "high": {
            "title": "High",
            "type": "integer"
          },
          "low": {
            "title": "Low",
            "type": "integer"
          },
          "medium": {
            "title": "Medium",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "fixable",
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "ContainerVulnerabilitySummary",
        "type": "object"
      },
...
      "FPPatternSchema": {
        "description": "False positive pattern schema.",
        "properties": {
          "container_name": {
            "title": "Container Name",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_matched": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Matched"
          },
          "match_count": {
            "title": "Match Count",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "start_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Line"
          }
        },
        "required": [
          "id",
          "container_name",
          "file_path",
          "rule_id",
          "reason",
          "created_by",
          "created_at",
          "match_count",
          "last_matched"
        ],
        "title": "FPPatternSchema",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Docker vulnerability scanner and remediation dashboard powered by Trivy",
    "title": "VulnForge",
    "version": "4.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/activity/container/{container_id}": {
      "get": {
        "description": "Get activities for a specific container.\n\nArgs:\n    container_id: Container ID\n    limit: Maximum number of activities to return\n    db: Database session\n\nReturns:\n    List of ActivityLog instances for the container",
        "operationId": "get_container_activities_api_v1_activity_container__container_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "container_id",
            "required": true,
            "schema": {
              "title": "Container Id",
              "type": "integer"
            }
          },
          {
            "description": "Maximum number of activities to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum number of activities to return",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityLog"
                  },
                  "title": "Response Get Container Activities Api V1 Activity Container  Container Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Container Activities",
        "tags": [
          "Activity"
        ]
      }
    },
...
    "/api/v1/containers/": {
      "get": {
        "description": "List all containers with optional pagination.",
        "operationId": "list_containers_api_v1_containers__get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContainerList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Containers",
        "tags": [
          "Containers"
        ]
      }
    },
    "/api/v1/containers/by-image": {
      "get": {
        "description": "Get containers matching an image name and optional tag.\n\nReturns a list because multiple containers may share the same image+tag\n(e.g. two containers running nginx:latest). Results are ordered by\nlast_scan_date DESC so the most recently scanned container is first.\n\nUsed by TideWatch for image-based vulnerability lookups.\n\nArgs:\n    image: Image repository name (e.g. \"nginx\", \"ghcr.io/org/app\").\n    tag: Optional image tag filter. If omitted, returns all tags.",
        "operationId": "get_containers_by_image_api_v1_containers_by_image_get",
        "parameters": [
          {
            "in": "query",
            "name": "image",
            "required": true,
            "schema": {
              "title": "Image",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Container"
                  },
                  "title": "Response Get Containers By Image Api V1 Containers By Image Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Containers By Image",
        "tags": [
          "Containers"
        ]
      }
    },
    "/api/v1/containers/by-name/{name}": {
      "get": {
        "description": "Get container by name (O(1) lookup).\n\nUsed by TideWatch for efficient container ID resolution without\nfetching the full container list.",
        "operationId": "get_container_by_name_api_v1_containers_by_name__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Container"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Container By Name",
        "tags": [
          "Containers"
        ]
      }
    },
...
    "/api/v1/containers/{container_id}": {
      "get": {
        "description": "Get container by ID.",
        "operationId": "get_container_api_v1_containers__container_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "container_id",
            "required": true,
            "schema": {
              "title": "Container Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Container"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Container",
        "tags": [
          "Containers"
        ]
      },
      "patch": {
        "description": "Update container fields (e.g., toggle is_my_project).",
        "operationId": "update_container_api_v1_containers__container_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "container_id",
            "required": true,
            "schema": {
              "title": "Container Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContainerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Container"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Container",
        "tags": [
          "Containers"
        ]
      }
    },
...
    "/api/v1/false-positive-patterns/container/{container_name}": {
      "get": {
        "description": "List FP patterns for a specific container.\n\nArgs:\n    container_name: Container name\n\nReturns:\n    List of FP patterns",
        "operationId": "list_container_fp_patterns_api_v1_false_positive_patterns_container__container_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "container_name",
            "required": true,
            "schema": {
              "title": "Container Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FPPatternSchema"
                  },
                  "title": "Response List Container Fp Patterns Api V1 False Positive Patterns Container  Container Name  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Container Fp Patterns",
        "tags": [
          "False Positive Patterns"
        ]
      }
    },
...
  }
}